cmake refactor work
