+++ src/core/model/CMakeLists.txt
@@ -1,0 +1,9 @@
+add_library(model_loader model_loader.cc)
+target_link_libraries(model_loader
+	geometry
+	mesh_pack
+	mesh_basic_material
+	mesh_flat_material
+    assimp
+)
+
