+++ src/core/model/model_loader.cc
@@ -3,7 +3,7 @@
 #include <assimp/scene.h>
 #include <assimp/Importer.hpp>
 
-#include "src/core/renderer/model_loader.h"
+#include "src/core/model/model_loader.h"
 #include "src/core/engine.h"
 #include "src/core/renderer/geometry.h"
 #include "src/core/elements/mesh_basic_material.h"
