+++ src/core/model/model_loader.h
