+++ src/core/scene/scene_manager.cc
@@ -1,5 +1,5 @@
 #include "src/core/scene/scene_manager.h"
-#include "src/core/renderer/model_loader.h"
+#include "src/core/model/model_loader.h"
 #include "src/core/engine.h"
 
 #include "src/core/elements/mesh_pack.h"
