+++ src/gui/imgui/CMakeLists.txt
@@ -1,0 +1,11 @@
+add_library(imgui_lib
+	gui_base.cc
+	gui_main.cc
+	gui.cc
+	object_inspector.cc
+	scene_explorer.cc
+	window.cc
+)
+target_link_libraries(imgui_lib
+	core
+)
